cross platform: fix stack overflow (ulong)
ulong (unsigned long) is 4 bytes on Windows while 8 bytes on *nix.

There is/are routines within ChakraCore assume casting from `double`
to `ulong` and then reacing to second item `[1]` is safe. However,
this ends up with stack overflow on *nix.

PAL has a definition for uint32 as (unsigned int) and it is already
widely used. Thus, replacing all the `ulong` to `uint32`

+++ lib/Common/Common/NumberUtilities.inl
@@ -42,21 +42,21 @@
 
 namespace Js
 {
-    NUMBER_UTIL_INLINE ulong &NumberUtilities::LuHiDbl(double &dbl)
+    NUMBER_UTIL_INLINE uint32 &NumberUtilities::LuHiDbl(double &dbl)
     {
 #if defined(__BIG_ENDIAN__)
-        return ((ulong *)&dbl)[0];
+        return ((uint32 *)&dbl)[0];
 #else //!BIG_ENDIAN
-        return ((ulong *)&dbl)[1];
+        return ((uint32 *)&dbl)[1];
 #endif //!BIG_ENDIAN
     }
 
-    NUMBER_UTIL_INLINE ulong &NumberUtilities::LuLoDbl(double &dbl)
+    NUMBER_UTIL_INLINE uint32 &NumberUtilities::LuLoDbl(double &dbl)
     {
 #if defined(__BIG_ENDIAN__)
-        return ((ulong *)&dbl)[1];
+        return ((uint32 *)&dbl)[1];
 #else //!BIG_ENDIAN
-        return ((ulong *)&dbl)[0];
+        return ((uint32 *)&dbl)[0];
 #endif //!BIG_ENDIAN
     }
